Refactor the functions
